backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
alembic==1.16.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.51.2
pinecone-client==3.3.1
httpx==0.27.2
python-multipart==0.0.9
celery==5.4.0
redis==5.2.1
pytest==8.3.3
black==24.8.0

